+++ src/Ext/Techno/Hooks.Misc.cpp
@@ -2,6 +2,22 @@
 
 #include <SpawnManagerClass.h>
 
+#pragma region SlaveManagerClass
+
+// Issue #601
+// Author : TwinkleStar
+DEFINE_HOOK(0x6B0C2C, SlaveManagerClass_FreeSlaves_SlavesFreeSound, 0x5)
+{
+	GET(TechnoClass*, pSlave, EDI);
+
+	auto pTypeExt = TechnoTypeExt::ExtMap.Find(pSlave->GetTechnoType());
+	int sound = pTypeExt->SlavesFreeSound.Get(RulesClass::Instance()->SlavesFreeSound);
+	if (sound != -1)
+		VocClass::PlayAt(sound, pSlave->Location);
+
+	return 0x6B0C65;
+}
+
 DEFINE_HOOK(0x6B0B9C, SlaveManagerClass_Killed_DecideOwner, 0x6)
 {
 	enum { KillTheSlave = 0x6B0BDF, ChangeSlaveOwner = 0x6B0BB4 };
@@ -38,6 +54,10 @@ DEFINE_HOOK(0x6B0B9C, SlaveManagerClass_Killed_DecideOwner, 0x6)
 DEFINE_PATCH(0x6B0BF7,
 	0x6A, 0x01  // push 1       // ignoreDefense=false->true
 );
+
+#pragma endregion
+
+#pragma region SpawnManagerClass
 
 DEFINE_HOOK(0x6B7265, SpawnManagerClass_AI_UpdateTimer, 0x6)
 {
@@ -99,6 +119,72 @@ DEFINE_HOOK(0x6B7600, SpawnManagerClass_AI_InitDestination, 0x6)
 	return R->Origin() == 0x6B7600 ? SkipGameCode1 : SkipGameCode2;
 }
 
+DEFINE_HOOK(0x6B7282, SpawnManagerClass_AI_PromoteSpawns, 0x5)
+{
+	GET(SpawnManagerClass*, pThis, ESI);
+
+	auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->Owner->GetTechnoType());
+	if (pTypeExt->Promote_IncludeSpawns)
+	{
+		for (auto i : pThis->SpawnedNodes)
+		{
+			if (i->Unit && i->Unit->Veterancy.Veterancy < pThis->Owner->Veterancy.Veterancy)
+				i->Unit->Veterancy.Add(pThis->Owner->Veterancy.Veterancy - i->Unit->Veterancy.Veterancy);
+		}
+	}
+
+	return 0;
+}
+
+#pragma endregion
+
+#pragma region WakeAnims
+
+namespace GrappleUpdateTemp
+{
+	TechnoClass* pThis;
+}
+
+DEFINE_HOOK(0x629E9B, ParasiteClass_GrappleUpdate_MakeWake_SetContext, 0x5)
+{
+	GET(TechnoClass*, pThis, ECX);
+	GrappleUpdateTemp::pThis = pThis;
+
+	return 0;
+}
+
+DEFINE_HOOK(0x629FA3, ParasiteClass_GrappleUpdate_MakeWake, 0x6)
+{
+	const auto pTypeExt = TechnoTypeExt::ExtMap.Find(GrappleUpdateTemp::pThis->GetTechnoType());
+	R->EDX(pTypeExt->Wake_Grapple.Get(pTypeExt->Wake.Get(RulesClass::Instance->Wake)));
+
+	return 0x629FA9;
+}
+
+DEFINE_HOOK(0x7365AD, UnitClass_Update_SinkingWake, 0x6)
+{
+	GET(UnitClass* const, pThis, ESI);
+
+	const auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->Type);
+	R->ECX(pTypeExt->Wake_Sinking.Get(pTypeExt->Wake.Get(RulesClass::Instance->Wake)));
+
+	return 0x7365B3;
+}
+
+DEFINE_HOOK(0x737F05, UnitClass_ReceiveDamage_SinkingWake, 0x6)
+{
+	GET(UnitClass* const, pThis, ESI);
+
+	const auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->Type);
+	R->ECX(pTypeExt->Wake_Sinking.Get(pTypeExt->Wake.Get(RulesClass::Instance->Wake)));
+
+	return 0x737F0B;
+}
+
+#pragma endregion
+
+#pragma region Misc
+
 // I must not regroup my forces.
 DEFINE_HOOK(0x739920, UnitClass_TryToDeploy_DisableRegroupAtNewConYard, 0x6)
 {
@@ -108,3 +194,5 @@ DEFINE_HOOK(0x739920, UnitClass_TryToDeploy_DisableRegroupAtNewConYard, 0x6)
 
 	return pRules->GatherWhenMCVDeploy ? DoNotSkipRegroup : SkipRegroup;
 }
+
+#pragma endregion
